Rename to simplify comparison
Rename the new vuex modules to make it easier to compare between
"pre" and "post" vuex store.

Feature: CMS-11254
