+++ src/MusicFestival.Vue.Template/Assets/Scripts/api/api.js
@@ -1,6 +1,6 @@
 /**
  * Wraps the calls to the ContentDeliveryAPI. It's used by the vuex
- * `contentModule` and the `ArtistContainerPage` component.
+ * `epiDataModel` module and the `ArtistContainerPage` component.
  */
 
 import axios from 'axios';
@@ -31,7 +31,7 @@ export default {
      * Getting data from ContentDeliveryAPI through regular routing (friendly
      * URLs) is enabled by the extensions in Infrastructure/ContentDeliveryExtendedRouting.
      * It is used in MusicFestival to get:
-     *  - page data, through the vuex `contentModule`
+     *  - page data, through the vuex `epiDataModel` module
      */
     getContentByFriendlyUrl: (friendlyUrl, parameters) => callRegularRouting('get', friendlyUrl, parameters),
