+++ src/MusicFestival.Vue.Template/Assets/Scripts/store/action-types.js
@@ -1,4 +1,4 @@
-// actions for the content module
+// actions for the epiDataModel module
 export const updateModelByFriendlyUrl = 'updateModelByFriendlyUrl';
 export const updateModelByContentLink = 'updateModelByContentLink';
 // actions for the main vuex store
