+++ src/MusicFestival.Vue.Template/Assets/Scripts/store/index.js
@@ -7,13 +7,13 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 
 // the module handling model state
-import content from './modules/contentModule.js';
+import epiDataModel from './modules/epiDataModel.js';
 
 // the module handling episerver specific state
-import epiContext from './modules/epiContextModule.js';
+import epiContext from './modules/epiContext.js';
 
 // the module handling app specific state
-import appContext from './modules/appContextModule.js';
+import appContext from './modules/appContext.js';
 
 import { UPDATE_URL } from './mutation-types.js';
 import { updateModelByFriendlyUrl, updateUrl } from './action-types.js';
@@ -23,7 +23,7 @@ Vue.use(Vuex);
 const store = new Vuex.Store({
     modules: {
         appContext,
-        content,
+        epiDataModel,
         epiContext
     },
     state: {
