+++ src/MusicFestival.Vue.Template/Assets/Scripts/store/modules/appContext.js
