+++ src/MusicFestival.Vue.Template/Assets/Scripts/store/modules/epiContext.js
