+++ src/MusicFestival.Vue.Template/Assets/Scripts/store/modules/epiDataModel.js
